Fixed tabs, replaced tools, deleted tmp files
