--- labs/12_i2c_oled_sprites/README.md~
@@ -1,64 +1,0 @@
-## Рисование спрайтов
-
-Лабораторная работа сделана в качестве ознакомления с инструментом **BMP to sprite conversion**.
-
-### Мотивация
-
-Когда в арсенале индикаторов появляется *черно-белый дисплей*, возникает множество идей как его можно использовать. Кто выводит серые *картинки*, скачанные или нарисованные самостоятельно, кто пишет удобный *интерфейс* для взаимодействия с МК, а кто создаёт *анимацию или игры*. В последнем случае не обходится без создания обликов различных персонажей и объектов. И один из способ нарисовать такого персонажа на экране - попиксельная отрисовка. Это способ достичь результат, но не самый быстрый и удобный способ.
-
-Пользование дисплеем с помощью МК **воодушевляет** на создание проектов. Поэтому реализация любой графической задумки должна быть **быстрой и удобной**. Иначе интерес пропадает. Причём потраченные временные ресурсы должны сводится к минимуму, обьём написанного кода не должнен выливаться за пределы памяти МК, а исполнение по возможности быстрое. Предсталенный здесь *гибкий инструмент* **BMP to sprite conversion** пригодится для реализации всевозможных идей. Начиная от создания небольших обьектов, персонажей и заканчивая большими движущимися фонами.
-
-### О программе
-
-Инструмент **BMP to sprite conversion** позволяет перевести рисунок из файла типа BMP в С код и дальше на дисплей. Состоит инструмент из двух частей. Первая часть - программа для преобразования BMP файла в код - **bmp2C**. Она находится в папке "labs/tools/BMP to C converter". Вторая часть - функции для вывода риуснка на экран. Эти функции включены в библиотеку управления дисплеем `oled_display` в этой лабораторной работе. Правила, по которым работает преобразование, можно найти в мануале "BMP to sprite converter manual", находящемся в той же папке рядом с **bmp2C**.
-
-### Быстрый старт
-1. Создайте картинку в формате BMP с *8 битной глубиной цвета*. Задайте имя картинке такое же, как имя самого спрайта. Нарисуйте небольшой пиксель-арт – *спрайт* – ту картинку, которую вы хотите вывести на экран. Картинка может быть трех «цветов»: белый, черный и прозрачный. Прозрачным будет считаться любой цвет кроме черного и белого. Сохраните картинку в отдельной папке, в которой больше нет других файлов.
-
-2. Запустите bmp2C программу, указав в качестве аргумента путь к папке с картинкой спрайта. В результате сгенерируется файл формата С, в котором содержится вся информация о вашем спрайте. Вторым аргументом к программе можно (но не обязательно) указать желаемое имя для создаваемого С файла. 
-3. Поместите сгенерированный файл в папку с вашим проектом. Убедитесь, что в проекте вы также используете `oled_display.c` и `oled_display.h` *с поддержкой рисования спрайтов*. Там этих файлах вы должны обнаружить упоминание спрайтов.
-
-4.	Нарисуем созданный спрайт на экране. В начале в `main.c` создадим переменную типа Sprite. Затем проинициализируем ее именно теми данными, которые хранятся в сгенерированном ранее файле. Перед этим портируем соответствующую функцию инициализации из того файла (extern). 
-```C
-#include “oled_display.h”
-<…>
-extern void init_dino_sprite(Sprite *sprite);
-Sprite dino;
-init_dino_sprite(&dino);
-```
-
-5.	Используем функцию рисования спрайта draw_sprite. Укажем аргументы: переменную типа `Sprite`, номер костюма (пока 0), координаты `x` и `y` на экране, считая от верхнего левого угла экрана до верхнего левого угла исходной картинки BMP. Опции пока оставим 0. 
-```C
-int16_t x = 10; int16_t y = 10;
-draw_sprite(dino, 0, x, y, 0);
-```
-
-6.	Запускаем и любуемся!
-
-<p align="center">
-  <img width="600" src="https://github.com/Levitsky-Ilya/stm32f0_ARM/blob/master/docs/images/sprite_on_screen.jpg" alt="sprite_on_screen.jpg"/>
-</p>
-
-Это был краткий экскурс, как рисовать спрайты. Далее идут тонкости, описанные в мануале "BMP to sprite converter manual".
-
-### Испытай себя
-
-1. Нарисуйте динозаврика.
-
-2. Сделайте много одинаковых спрайтов, используя одну картинку. Разбросайте их по экрану. Спрячьте некоторых немного за границей.
-
-3. Используйте костюмы: Сделайте экран загрузки или анимацию пламени.
-
-4. Используйте отражения: Сделайте пакмана, перемещающегося вправо-влево. Затем добавьте костюмы с движением вверх-вниз.
-
-5. Добавьте фон, нарисованный своими руками.
-
-6. Сделайте сразу несколько разных спрайтов - с костюмами или без.
-
-7. Вашего персонажа не всегда видно на любом фоне? Используйте опции вроде `DRAW_BLACK_INVERT`.
-
-8. Проверка на коллизию: пусть фон - лабиринт! Белый пакман коснулся белой стенки? - вернулся в начало.
-
-9. Сделайте мини-игру, используя по полной новый функционал!
-
-10. Функционала не хватает? Разберите весь код и улучшите/дополните его для себя! Если нововведение хорошо - оно может попасть в оригинальный код ;)
