+++ labs/12_i2c_oled_sprites/sprite_table.c
@@ -13,18 +13,18 @@ uint8_t dino_sprite_black_costume[448] = {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
 uint8_t dino_sprite_white_costume[448] = {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,248,12,4,7,1,1,241,241,241,241,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,3,6,4,12,248,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,255,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,255,248,8,8,8,8,248,0,0,0,0,0,0,0,0,0,0,0,0,0,0,192,64,64,64,112,16,31,0,0,0,0,0,0,0,0,0,0,0,0,224,32,35,35,35,35,35,35,35,35,63,1,1,1,1,1,1,255,0,0,0,0,1,1,1,15,8,56,32,32,48,16,16,30,2,2,2,3,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,225,129,129,129,129,255,0,0,0,0,0,0,0,0,0,0,0,63,32,224,128,128,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,192,64,64,127,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,7,4,4,4,4,252,128,128,0,0,0,0,0,0,0,0,0,0,128,0,0,0,0,0,0,0,224,32,60,4,4,7,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,255,0,0,0,0,48,48,48,254,2,2,3,2,2,254,0,0,0,0,63,32,32,224,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,3,2,2,2,2,2,2,2,3,0,0,0,0,0,3,2,2,2,2,2,2,2,3,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0};
 
 uint8_t *dino_sprite_black_costumes_bank[1] = {
-	(uint8_t*)(&dino_sprite_black_costume)
+    (uint8_t*)(&dino_sprite_black_costume)
 };
 uint8_t *dino_sprite_white_costumes_bank[1] = {
-	(uint8_t*)(&dino_sprite_white_costume)
+    (uint8_t*)(&dino_sprite_white_costume)
 };
 
 void init_dino_sprite(Sprite *sprite)
 {
-	sprite->black_costumes_bank = dino_sprite_black_costumes_bank;
-	sprite->white_costumes_bank = dino_sprite_white_costumes_bank;
+    sprite->black_costumes_bank = dino_sprite_black_costumes_bank;
+    sprite->white_costumes_bank = dino_sprite_white_costumes_bank;
 
-	sprite->width = 56; sprite->height = 8;
-	sprite->n_costumes = 1;
+    sprite->width = 56; sprite->height = 8;
+    sprite->n_costumes = 1;
 }
