+++ tools/bmp2C/BMP_to_sprite_converter_manual.pdf
